+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Remove `testutil.WaitFor`, use `testify.Eventually` helper if needed (#2920)
 - Remove testutil.WaitForPort, users can use testify.Eventually (#2926)
 - Rename `processorhelper.NewTraceProcessor` to `processorhelper.NewTracesProcessor` (#2935)
+- Rename `exporterhelper.NewTraceExporter` to `exporterhelper.NewTracesExporter` (#2937)
 
 ## 💡 Enhancements 💡